test: separate provider api contract integration coverage
